updated setUpZsh bash script

+++ setUpZsh.sh
@@ -1,12 +1,31 @@
 #!bin/sh
 echo; echo "Installing Zsh plugins..."; sleep 1
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-git clone https://github.com/zsh-users/zsh-completions ${ZSH_CUSTOM:-${ZSH:-~/.oh-my-zsh}/custom}/plugins/zsh-completions
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-
-echo; echo "Reloading Zsh..."; sleep 1;
-omz reload
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting" ]
+then    
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+  echo "Zsh syntax highlighting plugin installed!"
+fi
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-completions" ]
+then    
+  git clone https://github.com/zsh-users/zsh-completions ${ZSH_CUSTOM:-${ZSH:-~/.oh-my-zsh}/custom}/plugins/zsh-completions
+  echo "Zsh completions plugin installed!"
+fi
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions" ]
+then    
+  git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+  echo "Zsh autosuggestions plugin installed!"
+fi
 
 echo; echo "Copying zsh config files..."; sleep 1;
 cp config/.zshrc ~/.zshrc
-cp config/.zshlocalrc ~/.zshlocalrc
+echo ".zshrc file updated"
+if [ ! -f "$HOME/.zshlocalrc" ]
+then
+  cat ~/.zshlocalrc ~/.zshlocalrc_backup
+  cp config/.zshlocalrc ~/.zshlocalrc
+  echo ".zshlocalrc file created"
+fi
+echo "done!"
+
+echo; echo "Reloading Zsh..."; sleep 1;
+omz reload
